+++ src/com/zom/world/Pistol.java
@@ -10,7 +10,7 @@ public class Pistol extends Gun
   protected final static int FIRE_DELAY = 500;
   private long lastFireTime = 0;
 
-  public void fire(World w, int x, int y, int angle, int playerId)
+  public void fire(World w, int x, int y, byte angle, int playerId)
   {
     if (System.currentTimeMillis() - lastFireTime > FIRE_DELAY)
     {
